pyproject.toml
[tool.poetry]
name = "assignment3-nvidia-multimodalresearchassistant"
version = "0.1.0"
description = ""
authors = ["dharun-karthick-ramaraj <dharun.ramaraj@gmail.com>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "3.11"
bs4 = "^0.0.2"
requests = "^2.32.3"
selenium = "^4.25.0"
python-dotenv = "^1.0.1"
boto3 = "^1.35.45"
pandas = "^2.2.3"
snowflake-connector-python = "^3.12.2"
fastapi = "^0.115.3"
uvicorn = "^0.32.0"
streamlit = "^1.39.0"
st-clickable-images = "^0.0.3"
pymupdf = "^1.24.12"
pdfplumber = "^0.11.4"
openai = "^1.52.2"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
